librbd: refactor watcher types
