+++ src/librbd/MirroringWatcher.cc
@@ -6,6 +6,7 @@
 #include "include/rados/librados.hpp"
 #include "common/errno.h"
 #include "librbd/Utils.h"
+#include "librbd/watcher/Utils.h"
 
 #define dout_subsys ceph_subsys_rbd
 #undef dout_prefix
@@ -15,6 +16,8 @@ namespace librbd {
 
 using namespace mirroring_watcher;
 using namespace watcher;
+
+using librbd::util::create_rados_callback;
 
 namespace {
 
@@ -46,7 +49,7 @@ void MirroringWatcher<I>::notify_mode_updated(librados::IoCtx &io_ctx,
   bufferlist bl;
   ::encode(NotifyMessage{ModeUpdatedPayload{mirror_mode}}, bl);
 
-  librados::AioCompletion *comp = util::create_rados_callback(on_finish);
+  librados::AioCompletion *comp = create_rados_callback(on_finish);
   int r = io_ctx.aio_notify(RBD_MIRRORING, comp, bl, NOTIFY_TIMEOUT_MS,
                             nullptr);
   assert(r == 0);
@@ -76,7 +79,7 @@ void MirroringWatcher<I>::notify_image_updated(
   ::encode(NotifyMessage{ImageUpdatedPayload{
       mirror_image_state, image_id, global_image_id}}, bl);
 
-  librados::AioCompletion *comp = util::create_rados_callback(on_finish);
+  librados::AioCompletion *comp = create_rados_callback(on_finish);
   int r = io_ctx.aio_notify(RBD_MIRRORING, comp, bl, NOTIFY_TIMEOUT_MS,
                             nullptr);
   assert(r == 0);
@@ -104,9 +107,8 @@ void MirroringWatcher<I>::handle_notify(uint64_t notify_id, uint64_t handle,
     return;
   }
 
-  apply_visitor(HandlePayloadVisitor<MirroringWatcher<I>>(this, notify_id,
-                                                          handle),
-                notify_message.payload);
+  apply_visitor(watcher::util::HandlePayloadVisitor<MirroringWatcher<I>>(
+                  this, notify_id, handle), notify_message.payload);
 }
 
 template <typename I>
