+++ src/librbd/watcher/Types.h
@@ -4,13 +4,11 @@
 #ifndef CEPH_LIBRBD_WATCHER_TYPES_H
 #define CEPH_LIBRBD_WATCHER_TYPES_H
 
+#include "include/int_types.h"
 #include "include/buffer_fwd.h"
 #include "include/encoding.h"
-#include "include/Context.h"
 
-namespace ceph {
-class Formatter;
-}
+namespace ceph { class Formatter; }
 
 namespace librbd {
 
@@ -18,65 +16,34 @@ class Watcher;
 
 namespace watcher {
 
-struct C_NotifyAck : public Context {
-  Watcher *watcher;
-  CephContext *cct;
-  uint64_t notify_id;
-  uint64_t handle;
-  bufferlist out;
-
-  C_NotifyAck(Watcher *watcher, uint64_t notify_id, uint64_t handle);
-  void finish(int r) override;
-};
-
-template <typename Watcher>
-struct HandlePayloadVisitor : public boost::static_visitor<void> {
-  Watcher *watcher;
-  uint64_t notify_id;
+struct ClientId {
+  uint64_t gid;
   uint64_t handle;
 
-  HandlePayloadVisitor(Watcher *watcher_, uint64_t notify_id_,
-      uint64_t handle_)
-    : watcher(watcher_), notify_id(notify_id_), handle(handle_)
-  {
+  ClientId() : gid(0), handle(0) {}
+  ClientId(uint64_t gid, uint64_t handle) : gid(gid), handle(handle) {}
+
+  void encode(bufferlist& bl) const;
+  void decode(bufferlist::iterator& it);
+  void dump(Formatter *f) const;
+
+  inline bool is_valid() const {
+    return (*this != ClientId());
   }
 
-  template <typename P>
-  inline void operator()(const P &payload) const {
-    C_NotifyAck *ctx = new C_NotifyAck(watcher, notify_id, handle);
-    if (watcher->handle_payload(payload, ctx)) {
-      ctx->complete(0);
+  inline bool operator==(const ClientId &rhs) const {
+    return (gid == rhs.gid && handle == rhs.handle);
+  }
+  inline bool operator!=(const ClientId &rhs) const {
+    return !(*this == rhs);
+  }
+  inline bool operator<(const ClientId &rhs) const {
+    if (gid != rhs.gid) {
+      return gid < rhs.gid;
+    } else {
+      return handle < rhs.handle;
     }
   }
-};
-
-class EncodePayloadVisitor : public boost::static_visitor<void> {
-public:
-  explicit EncodePayloadVisitor(bufferlist &bl) : m_bl(bl) {}
-
-  template <typename P>
-  inline void operator()(const P &payload) const {
-    ::encode(static_cast<uint32_t>(P::NOTIFY_OP), m_bl);
-    payload.encode(m_bl);
-  }
-
-private:
-  bufferlist &m_bl;
-};
-
-class DecodePayloadVisitor : public boost::static_visitor<void> {
-public:
-  DecodePayloadVisitor(__u8 version, bufferlist::iterator &iter)
-    : m_version(version), m_iter(iter) {}
-
-  template <typename P>
-  inline void operator()(P &payload) const {
-    payload.decode(m_version, m_iter);
-  }
-
-private:
-  __u8 m_version;
-  bufferlist::iterator &m_iter;
 };
 
 template <typename ImageCtxT>
@@ -86,5 +53,10 @@ struct Traits {
 
 } // namespace watcher
 } // namespace librbd
+
+std::ostream &operator<<(std::ostream &out,
+                         const librbd::watcher::ClientId &client);
+
+WRITE_CLASS_ENCODER(librbd::watcher::ClientId);
 
 #endif // CEPH_LIBRBD_WATCHER_TYPES_H
